// src/app/interfaces/kcdb-comparisons.interface.ts
// Generated by https://quicktype.io

export interface SubFieldAM {
  id:              string;
  label:           string;
  value:           string;
  additionalValue: null;
  locked:          boolean;
}


// Generated by https://quicktype.io

export interface Comparisons {
  recordsFiltered: number;
  allCmcs:         number[];
  data:            Datum[];
  draw:            number;
  recordsTotal:    number;
}

export interface Datum {
  id:                   number;
  comparisonIdentifier: string;
  description:          string;
  startYear:            string;
  endYear:              null | string;
  metrologyAreaName:    string;
  subField:             string;
  parameters:           string;
  status:               string;
  instituteName:        string;
  instituteCode:        string;
  comparisonType:       string;
  oneYearDefined:       boolean;
}

export interface ParamasSearchComparisons{

  draw:           string
  page:           number
  start:          string
  length:         string
  size:           number
  sort:           string
  areaId:         string
  subfieldId:     string
  comparisonType: string
  organizationId: string
  startYear:      string
  endYear:        string
  keywords:       string
  _:              string

}